Do not render experience rating banner if Ads account is disconnected.
