+++ js/src/components/experience-rating-banner/constants.js
@@ -1,0 +1,1 @@
+export const BANNER_DISMISSED_KEY = 'experience-rating-banner-dismissed';
